:pray: pave the path towards ReturnNan default argminmax for floats
