+++ src/simd/mod.rs
@@ -7,11 +7,11 @@ mod generic;
 pub use generic::*;
 // FLOAT
 mod simd_f16_ignore_nans;
-// mod simd_f16_return_nans;
+// mod simd_f16_return_nans; // TODO: not supported yet
 mod simd_f32_ignore_nans;
-mod simd_f32_return_nans;
+mod simd_f32;
 mod simd_f64_ignore_nans;
-mod simd_f64_return_nans;
+mod simd_f64;
 // SIGNED INT
 mod simd_i16;
 mod simd_i32;
